ESP32 CI fixes

+++ Adafruit_TCS34725.h
@@ -210,7 +210,7 @@ public:
   uint16_t calculateColorTemperature_dn40(uint16_t r, uint16_t g, uint16_t b,
                                           uint16_t c);
   uint16_t calculateLux(uint16_t r, uint16_t g, uint16_t b);
-  void write8(uint8_t reg, uint32_t value);
+  void write8(uint8_t reg, uint8_t value);
   uint8_t read8(uint8_t reg);
   uint16_t read16(uint8_t reg);
   void setInterrupt(boolean flag);
